Move slipping sound to the server
