+++ Content.Server/GameObjects/Components/Movement/SlipperyComponent.cs
@@ -1,5 +1,10 @@
-﻿﻿using Content.Shared.GameObjects.Components.Movement;
+﻿using Content.Shared.Audio;
+using Content.Shared.GameObjects.Components.Movement;
+using Robust.Server.GameObjects.EntitySystems;
 using Robust.Shared.GameObjects;
+using Robust.Shared.GameObjects.Systems;
+using Robust.Shared.Serialization;
+using Robust.Shared.ViewVariables;
 
 namespace Content.Server.GameObjects.Components.Movement
 {
@@ -7,9 +12,26 @@ namespace Content.Server.GameObjects.Components.Movement
     [ComponentReference(typeof(SharedSlipperyComponent))]
     public class SlipperyComponent : SharedSlipperyComponent
     {
-        public override ComponentState GetComponentState()
+        /// <summary>
+        ///     Path to the sound to be played when a mob slips.
+        /// </summary>
+        [ViewVariables]
+        private string SlipSound { get; set; } = "/Audio/Effects/slip.ogg";
+
+        public override void ExposeData(ObjectSerializer serializer)
         {
-            return base.GetComponentState(); // TODO
+            base.ExposeData(serializer);
+
+            serializer.DataField(this, x => SlipSound, "slipSound", "/Audio/Effects/slip.ogg");
+        }
+
+        protected override void OnSlip()
+        {
+            if (!string.IsNullOrEmpty(SlipSound))
+            {
+                EntitySystem.Get<AudioSystem>()
+                    .PlayFromEntity(SlipSound, Owner, AudioHelpers.WithVariation(0.2f));
+            }
         }
     }
 }
