+++ Content.Shared/Content.Shared.csproj
@@ -29,6 +29,5 @@
     <Folder Include="GameObjects\Components\Trigger\" />
     <EmbeddedResource Include="Text\Names\*.txt" />
     <Folder Include="GameObjects\Components\Visualizers" />
-    <Folder Include="Mobs" />
   </ItemGroup>
 </Project>
