+++ Content.Shared/GameObjects/Components/Movement/SharedSlipperyComponent.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using Content.Shared.GameObjects.Components.Mobs;
 using Content.Shared.GameObjects.EntitySystems;
 using Content.Shared.Physics;
@@ -24,7 +25,7 @@ namespace Content.Shared.GameObjects.Components.Movement
         ///     The list of entities that have been slipped by this component,
         ///     and which have not stopped colliding with its owner yet.
         /// </summary>
-        private readonly List<EntityUid> _slipped = new List<EntityUid>();
+        protected readonly List<EntityUid> _slipped = new List<EntityUid>();
 
         /// <summary>
         ///     How many seconds the mob will be paralyzed for.
